Update behavior_RF.m
Corrected the way that the variable behaveblock is created. behaveblock is the list of runs in Tosca, ordered sequentially. The previous version did not work if the number of runs was over 9 because  MATLAB alphabetizes Run 1, Run 10, Run 2, etc instead of Run 1... Run 9, Run 10.  Added code to make sure the runs were ordered correctly.